Renamed task_worker.py as worker.py
